+++ MacPackerTests/ArchiveHandlerXzTests.swift
@@ -1,17 +1,17 @@
 //
-//  ArchiveHandlerZipTests.swift
+//  ArchiveHandlerXzTests.swift
 //  MacPacker
 //
-//  Created by Stephan Arenswald on 05.09.25.
+//  Created by Stephan Arenswald on 16.09.25.
 //
-
 
 import Foundation
 import XCTest
 @testable import MacPacker
 
-final class ArchiveHandlerZipTests: XCTestCase {
+final class ArchiveHandlerXzTests: XCTestCase {
     let testBase: ArchiveTestBase = ArchiveTestBase()
+    let defaultTestFile: String = "defaultArchive.tar.xz"
     
     override class func setUp() {
         super.setUp()
@@ -23,44 +23,57 @@ final class ArchiveHandlerZipTests: XCTestCase {
     }
     
     func testLoadFile() throws {
-        let tf = try testBase.getTestFile(name: "archive.zip")
+        let tf = try testBase.getTestFile(name: defaultTestFile)
         
         let archive = try Archive2(url: tf)
         let count = archive.items.count
-        XCTAssertEqual(count, 2)
+        XCTAssertEqual(count, 1)
     }
     
     func testExtractFile() throws {
-        let tf = try testBase.getTestFile(name: "archive.zip")
+        let tf = try testBase.getTestFile(name: defaultTestFile)
         
         let archive = try Archive2(url: tf)
-        let url = archive.extractFileToTemp(archive.items[0])
-        XCTAssertNotNil(url)
         
-        let content = try String(contentsOf: url!, encoding: .utf8)
-        XCTAssertEqual(content, "James Bond.\n")
-    }
-    
-    func testExtractNestedFile() throws {
-        let tf = try testBase.getTestFile(name: "archiveNested1.zip")
-        
-        let archive = try Archive2(url: tf)
         XCTAssertEqual(try archive.open(archive.items[0]), .success)
         if let se = archive.currentStackEntry {
-            XCTAssertEqual(se.name, "Folder")
+            XCTAssertEqual(se.name, "defaultArchive.tar")
         }
 
         XCTAssertEqual(try archive.open(archive.items[1]), .success)
         if let se = archive.currentStackEntry {
-            XCTAssertEqual(se.name, "archive.tar.lz4")
+            XCTAssertEqual(se.name, "defaultContent")
         }
+
+        let url = archive.extractFileToTemp(archive.items[3])
+        XCTAssertNotNil(url)
+        
+        let content = try String(contentsOf: url!, encoding: .utf8)
+        XCTAssertEqual(content, "Hello World!\n")
+    }
+    
+    func testNested() throws {
+        let tf = try testBase.getTestFile(name: defaultTestFile)
+        
+        let archive = try Archive2(url: tf)
+        XCTAssertEqual(try archive.open(archive.items[0]), .success)
+        if let se = archive.currentStackEntry {
+            XCTAssertEqual(se.name, "defaultArchive.tar")
+        }
+
         XCTAssertEqual(try archive.open(archive.items[1]), .success)
         if let se = archive.currentStackEntry {
-            XCTAssertEqual(se.name, "archive.tar")
+            XCTAssertEqual(se.name, "defaultContent")
         }
+
         XCTAssertEqual(try archive.open(archive.items[1]), .success)
         if let se = archive.currentStackEntry {
-            XCTAssertEqual(se.name, "archive.tar")
+            XCTAssertEqual(se.name, "folder")
+        }
+        
+        XCTAssertEqual(try archive.open(archive.items[1]), .success)
+        if let se = archive.currentStackEntry {
+            XCTAssertEqual(se.name, "archive.zip")
         }
         
         let url = archive.extractFileToTemp(archive.items[1])
@@ -73,14 +86,13 @@ final class ArchiveHandlerZipTests: XCTestCase {
     }
     
     func testExtractFullArchive() throws {
-        let archive = try testBase.getArchiveFor(name: "archive.zip")
+        let archive = try testBase.getArchiveFor(name: defaultTestFile)
         let service = ArchiveService()
         service.extract(
             archive: archive,
             to: ArchiveTestBase.tempDirectoryURL)
         print("Extracted to \(ArchiveTestBase.tempDirectoryURL)")
         
-        XCTAssertTrue(testBase.fileExistsInTemp("bond.txt"))
-        XCTAssertTrue(testBase.fileExistsInTemp("hello.txt"))
+        XCTAssertTrue(testBase.fileExistsInTemp("defaultArchive.tar"))
     }
 }