+++ MacPacker/CHANGELOG.md
@@ -1,5 +1,6 @@
 v0.7
 - feat: lzh, lha, lzx support
+- feat: close internal previewer using Space, or Esc
 - core: new architecture for archive handlers
 
 v0.6